Changed all threshold index variable types from unsigned to size_t in the monte_carlo_collision_core and monte_carlo_collision_photon package. Changed all trials counter variable types from unsigned to DistributionTraits::Counter in the monte_carlo_collision_photon package.

+++ packages/monte_carlo/collision/core/src/MonteCarlo_AtomicReaction.hpp
@@ -43,7 +43,7 @@ public:
 
   //! Return the cross section at the given energy (efficient)
   virtual double getCrossSection( const double energy,
-                                  const unsigned bin_index ) const = 0;
+                                  const size_t bin_index ) const = 0;
 
   //! Return the threshold energy
   virtual double getThresholdEnergy() const = 0;
