+++ packages/monte_carlo/collision/photon/src/MonteCarlo_CoherentPhotoatomicReaction.hpp
@@ -29,7 +29,7 @@ public:
   CoherentPhotoatomicReaction(
        const std::shared_ptr<const std::vector<double> >& incoming_energy_grid,
        const std::shared_ptr<const std::vector<double> >& cross_section,
-       const unsigned threshold_energy_index,
+       const size_t threshold_energy_index,
        const std::shared_ptr<const CoherentScatteringDistribution>&
        scattering_distribution );
 
@@ -37,7 +37,7 @@ public:
   CoherentPhotoatomicReaction(
     const std::shared_ptr<const std::vector<double> >& incoming_energy_grid,
     const std::shared_ptr<const std::vector<double> >& cross_section,
-    const unsigned threshold_energy_index,
+    const size_t threshold_energy_index,
     const std::shared_ptr<const Utility::HashBasedGridSearcher<double> >&
     grid_searcher,
     const std::shared_ptr<const CoherentScatteringDistribution>&
