+++ packages/monte_carlo/collision/photon/src/MonteCarlo_CrossSectionBasedPhotonuclearProductionReaction_def.hpp
@@ -20,7 +20,7 @@ CrossSectionBasedPhotonuclearProductionReaction<OutgoingParticleType>::CrossSect
        const PhotonuclearReactionType reaction_type,
        const std::vector<unsigned>& photon_production_ids,
        const double q_value,
-       const unsigned threshold_energy_index,
+       const size_t threshold_energy_index,
        const std::shared_ptr<const std::vector<double> >& incoming_energy_grid,
        const std::shared_ptr<const std::vector<double> >& cross_section,
        const std::vector<std::shared_ptr<const NuclearScatteringDistribution<PhotonState,OutgoingParticleType> > >&
@@ -64,7 +64,7 @@ void CrossSectionBasedPhotonuclearProductionReaction<OutgoingParticleType>::reac
   testPrecondition( photon.getEnergy() > 0.0 );
 
   // Create the additional particles
-  for( unsigned i = 0; i < d_photon_production_ids.size(); ++i )
+  for( size_t i = 0; i < d_photon_production_ids.size(); ++i )
   {
     std::shared_ptr<OutgoingParticleType> new_particle(
 			      new OutgoingParticleType( photon, true, true ) );
