+++ packages/monte_carlo/collision/photon/src/MonteCarlo_IncoherentPhotoatomicReaction_def.hpp
@@ -21,7 +21,7 @@ template<typename InterpPolicy, bool processed_cross_section>
 IncoherentPhotoatomicReaction<InterpPolicy,processed_cross_section>::IncoherentPhotoatomicReaction(
               const std::shared_ptr<const std::vector<double> >& incoming_energy_grid,
 	      const std::shared_ptr<const std::vector<double> >& cross_section,
-	      const unsigned threshold_energy_index,
+	      const size_t threshold_energy_index,
               const std::shared_ptr<const IncoherentPhotonScatteringDistribution>&
 	      scattering_distribution )
   : BaseType( incoming_energy_grid,
@@ -38,7 +38,7 @@ template<typename InterpPolicy, bool processed_cross_section>
 IncoherentPhotoatomicReaction<InterpPolicy,processed_cross_section>::IncoherentPhotoatomicReaction(
        const std::shared_ptr<const std::vector<double> >& incoming_energy_grid,
        const std::shared_ptr<const std::vector<double> >& cross_section,
-       const unsigned threshold_energy_index,
+       const size_t threshold_energy_index,
        const std::shared_ptr<const Utility::HashBasedGridSearcher<double> >&
        grid_searcher,
        const std::shared_ptr<const IncoherentPhotonScatteringDistribution>&
