+++ packages/monte_carlo/collision/photon/src/MonteCarlo_SubshellIncoherentAdjointPhotoatomicReaction.hpp
@@ -34,7 +34,7 @@ public:
   SubshellIncoherentAdjointPhotoatomicReaction(
        const std::shared_ptr<const std::vector<double> >& incoming_energy_grid,
        const std::shared_ptr<const std::vector<double> >& cross_section,
-       const unsigned threshold_energy_index,
+       const size_t threshold_energy_index,
        const std::shared_ptr<SubshellIncoherentAdjointPhotonScatteringDistribution>&
        scattering_distribution );
 
@@ -42,7 +42,7 @@ public:
   SubshellIncoherentAdjointPhotoatomicReaction(
        const std::shared_ptr<const std::vector<double> >& incoming_energy_grid,
        const std::shared_ptr<const std::vector<double> >& cross_section,
-       const unsigned threshold_energy_index,
+       const size_t threshold_energy_index,
        const std::shared_ptr<const Utility::HashBasedGridSearcher<double> >&
        grid_searcher,
        const std::shared_ptr<SubshellIncoherentAdjointPhotonScatteringDistribution>&
@@ -57,7 +57,7 @@ public:
 
   //! Return the cross section at the given energy (efficient)
   double getCrossSection( const double energy,
-                          const unsigned bin_index ) const override;
+                          const size_t bin_index ) const override;
   
   //! Return the reaction type
   virtual AdjointPhotoatomicReactionType getReactionType() const override;
