+++ packages/monte_carlo/collision/photon/src/MonteCarlo_SubshellPhotoelectricPhotoatomicReaction.hpp
@@ -25,7 +25,7 @@ class SubshellPhotoelectricPhotoatomicReaction : public PhotoelectricPhotoatomic
   SubshellPhotoelectricPhotoatomicReaction(
        const std::shared_ptr<const std::vector<double> >& incoming_energy_grid,
        const std::shared_ptr<const std::vector<double> >& cross_section,
-       const unsigned threshold_energy_index,
+       const size_t threshold_energy_index,
        const Data::SubshellType interaction_subshell,
        const double binding_energy );
 
@@ -33,7 +33,7 @@ class SubshellPhotoelectricPhotoatomicReaction : public PhotoelectricPhotoatomic
   SubshellPhotoelectricPhotoatomicReaction(
     const std::shared_ptr<const std::vector<double> >& incoming_energy_grid,
     const std::shared_ptr<const std::vector<double> >& cross_section,
-    const unsigned threshold_energy_index,
+    const size_t threshold_energy_index,
     const std::shared_ptr<const Utility::HashBasedGridSearcher<double> >&
     grid_searcher,
     const Data::SubshellType interaction_subshell,
